chore(release): 1.0.3

+++ core/src/main/java/dev/wearkit/core/engine/Chrono.java
@@ -1,6 +1,6 @@
 package dev.wearkit.core.engine;
 
 interface Chrono {
-    float getStartTime();
-    float getElapsedTime();
+    double getStartTime();
+    double getElapsedTime();
 }
